fix: #50 첫방문 환영 링크로 이동
